Reorganized page details folders
- Moved pageElements and pageConditions into pageDetails because they are
  technically subsections of pageDetails layout.
- Converted pageElements and pageCondition layout views into composite
  view to clean up their child collection views
