+++ src-backbone/app/css/modules/builderPageDetails.less
@@ -121,7 +121,7 @@
     }
 }
 
-div#elements-list ul{
+ul#elements-list{
     list-style: none;
     margin: 0;
     padding: 0;
