+++ src-backbone/app/js/templates/builder/pageDetails/pageConditions/showIfsCompositeView.hbs
@@ -1,6 +1,6 @@
 <h1 class="checkbox-heading">
     {{ i18n "Show this page conditionally?" }}
-    <input type="checkbox" id="show-page-conditionally" {{#if showPageConditionally }}checked{{/if}}>
+    <input type="checkbox" id="show-page-conditionally" {{#if pageHasConditions }}checked{{/if}}>
 </h1>
 
 <div id="condition-editor"></div>
