+++ src-backbone/app/js/templates/builder/pageDetails/pageDetailsLayoutView.hbs
@@ -1,2 +1,2 @@
-<section id="conditions" class="details-panel"></section>
+<section id="show-ifs" class="details-panel"></section>
 <section id="elements" class="details-panel"></section>
