+++ src-backbone/app/js/templates/builder/pageDetails/pageElements/choices/choicesItemView.hbs
