+++ src-backbone/app/js/templates/builder/pageDetails/pageElements/elementsCompositeView.hbs
@@ -1,6 +1,6 @@
 <h1>{{ i18n "Elements on this page" }}</h1>
 
-<div id="elements-list"></div>
+<ul id="elements-list"></ul>
 
 <h1>{{ i18n "Create a new element" }}</h1>
