+++ src-backbone/app/js/templates/builder/pageDetails/pageElements/types/choicesView.hbs
