+++ src-backbone/app/js/templates/builder/pageDetails/pageElements/types/pluginEntryView.hbs
