+++ src-backbone/app/js/templates/builder/pageDetails/pageElements/types/pluginView.hbs
