--- src-backbone/app/js/views/builder/pageConditions/showIfsCollectionView.js
@@ -1,7 +1,0 @@
-module.exports = Marionette.CollectionView.extend({
-
-    tagName: 'div',
-
-    childView: require('./showIfItemView'),
-
-});
