+++ src-backbone/app/js/views/builder/pageDetails/pageElements/choices/choicesCollectionView.js
