+++ src-backbone/app/js/views/builder/pageDetails/pageElements/emptyElementsView.js
@@ -1,6 +1,6 @@
 module.exports = Marionette.ItemView.extend({
 
-    template: require('templates/builder/pageElements/emptyElementsView'),
+    template: require('templates/builder/pageDetails/pageElements/emptyElementsView'),
     tagName: 'li',
     className: 'empty-elements',
