+++ src-backbone/app/js/views/builder/pageDetails/pageElements/types/multiSelectView.js
@@ -1,7 +1,7 @@
 module.exports = Marionette.LayoutView.extend({
 
-    template: require('templates/builder/pageElements/types/choicesView'),
-    className: 'has-choices radio-select', // To avoid conflicting with BootStrap's 'radio' class
+    template: require('templates/builder/pageDetails/pageElements/types/choicesView'),
+    className: 'has-choices multi-select',
 
     regions: {
         choicesList: 'div.choices-list',
