+++ src-backbone/app/js/views/builder/pageDetails/pageElements/types/pictureView.js
