+++ src-backbone/app/js/views/builder/pageDetails/pageElements/types/radioView.js
@@ -1,0 +1,14 @@
+module.exports = Marionette.LayoutView.extend({
+
+    template: require('templates/builder/pageDetails/pageElements/types/choicesView'),
+    className: 'has-choices radio-select', // To avoid conflicting with BootStrap's 'radio' class
+
+    regions: {
+        choicesList: 'div.choices-list',
+    },
+
+    behaviors: {
+        ChoiceBasedElementBehavior: { }
+    },
+
+});
