--- src-backbone/app/js/views/builder/pageElements/elementsCollectionView.js
@@ -1,13 +1,0 @@
-let EmptyView = require('./emptyElementsView');
-let ItemView = require('./elementItemView');
-
-
-module.exports = Marionette.CollectionView.extend({
-
-    tagName: 'ul',
-
-    emptyView: EmptyView,
-
-    childView: ItemView,
-
-});
