--- src-backbone/app/js/views/builder/pageElements/types/selectView.js
@@ -1,14 +1,0 @@
-module.exports = Marionette.LayoutView.extend({
-
-    template: require('templates/builder/pageElements/types/choicesView'),
-    className: 'has-choices select',
-
-    regions: {
-        choicesList: 'div.choices-list',
-    },
-
-    behaviors: {
-        ChoiceBasedElementBehavior: { }
-    },
-
-});
